Revert "estructura mejorada"
