+++ src/app/static/images/registro-image.jpg
