+++ src/app/templates/base.html
@@ -14,6 +14,6 @@
 <body>
     {% block body %}      {% endblock %}
 
-    <script src="./static/js/burger-menu.js"></script>
+    <script src="../../js/burger-menu.js"></script>
 </body>
 </html>
