--- src/backend/views/_init__.py
@@ -1,1 +1,0 @@
-from .calcularFecha import show_class_per_date
